Console representation constraints overhaul
